Remove references to selectInput_b and radioInput_b
