+++ src/signals/incident-management/components/RadioInput/index.test.js
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MPL-2.0
-// Copyright (C) 2018 - 2021 Gemeente Amsterdam
+// Copyright (C) 2018 - 2022 Gemeente Amsterdam
 import { render } from '@testing-library/react'
 import { withAppContext } from 'test/utils'
 import priorityList from 'signals/incident-management/definitions/priorityList'
@@ -7,18 +7,14 @@ import priorityList from 'signals/incident-management/definitions/priorityList'
 import RadioInput from '.'
 
 describe('<RadioInput />', () => {
-  const handler = () => ({ type: 'radio' })
   const props = {
     name: 'priority',
     display: 'Urgentie',
     values: priorityList,
   }
-  const RadioInputRender = RadioInput(props)
 
   it('should render a list of radio buttons', () => {
-    const { container } = render(
-      withAppContext(<RadioInputRender handler={handler} />)
-    )
+    const { container } = render(withAppContext(<RadioInput {...props} />))
 
     expect(container.querySelectorAll('input[type=radio]')).toHaveLength(
       priorityList.length
@@ -28,9 +24,7 @@ describe('<RadioInput />', () => {
   it('should display info text', () => {
     const currentValue = priorityList[0]
     const { getByText } = render(
-      withAppContext(
-        <RadioInputRender handler={handler} value={currentValue.key} />
-      )
+      withAppContext(<RadioInput {...props} value={currentValue.key} />)
     )
 
     expect(getByText(new RegExp(currentValue.info))).toBeInTheDocument()
@@ -38,13 +32,13 @@ describe('<RadioInput />', () => {
 
   it('should not display info text', () => {
     const currentValue = priorityList[0]
-    const RenderFunc = RadioInput({
+    const props = {
       name: 'priority',
       display: 'Urgentie',
-    })
+    }
 
     const { queryByText } = render(
-      withAppContext(<RenderFunc handler={handler} value={currentValue.key} />)
+      withAppContext(<RadioInput {...props} value={currentValue.key} />)
     )
 
     expect(queryByText(new RegExp(currentValue.info))).not.toBeInTheDocument()
@@ -52,7 +46,7 @@ describe('<RadioInput />', () => {
 
   it('should display a label', () => {
     const { container, rerender } = render(
-      withAppContext(<RadioInputRender handler={handler} />)
+      withAppContext(<RadioInput {...props} />)
     )
 
     expect(
@@ -60,8 +54,7 @@ describe('<RadioInput />', () => {
     ).toHaveLength(1)
 
     const propsWithoutDisplay = { ...props, display: undefined }
-    const Render = RadioInput(propsWithoutDisplay)
-    rerender(withAppContext(<Render handler={handler} />))
+    rerender(withAppContext(<RadioInput {...propsWithoutDisplay} />))
 
     expect(
       container.querySelectorAll(`label[for="form${props.name}"]`)
